Combine V3 and V4
