--- src/Microsoft.Health.Dicom.SqlServer/Features/ExtendedQueryTag/SqlExtendedQueryTagStoreV4.cs
@@ -1,97 +1,0 @@
-﻿// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-using System.Collections.Generic;
-using System.Data;
-using System.Globalization;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Data.SqlClient;
-using Microsoft.Extensions.Logging;
-using Microsoft.Health.Dicom.Core.Exceptions;
-using Microsoft.Health.Dicom.Core.Features.ExtendedQueryTag;
-using Microsoft.Health.Dicom.SqlServer.Features.Schema;
-using Microsoft.Health.Dicom.SqlServer.Features.Schema.Model;
-using Microsoft.Health.SqlServer.Features.Client;
-using Microsoft.Health.SqlServer.Features.Storage;
-
-namespace Microsoft.Health.Dicom.SqlServer.Features.ExtendedQueryTag
-{
-    internal class SqlExtendedQueryTagStoreV4 : SqlExtendedQueryTagStoreV3
-    {
-        public SqlExtendedQueryTagStoreV4(
-           SqlConnectionWrapperFactory sqlConnectionWrapperFactory,
-           ILogger<SqlExtendedQueryTagStoreV4> logger)
-            : base(sqlConnectionWrapperFactory, logger)
-        {
-        }
-
-        public override SchemaVersion Version => SchemaVersion.V4;
-
-        public override async Task<IReadOnlyList<int>> AddExtendedQueryTagsAsync(IEnumerable<AddExtendedQueryTagEntry> extendedQueryTagEntries, int maxAllowedCount, CancellationToken cancellationToken)
-        {
-            using SqlConnectionWrapper sqlConnectionWrapper = await ConnectionWrapperFactory.ObtainSqlConnectionWrapperAsync(cancellationToken);
-            using SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateSqlCommand();
-
-            IEnumerable<AddExtendedQueryTagsInputTableTypeV1Row> rows = extendedQueryTagEntries.Select(ToAddExtendedQueryTagsInputTableTypeV1Row);
-            VLatest.AddExtendedQueryTags.PopulateCommand(sqlCommandWrapper, maxAllowedCount, new VLatest.AddExtendedQueryTagsTableValuedParameters(rows));
-
-            try
-            {
-                var keys = new List<int>();
-                using SqlDataReader reader = await sqlCommandWrapper.ExecuteReaderAsync(cancellationToken);
-                while (await reader.ReadAsync(cancellationToken))
-                {
-                    keys.Add(reader.ReadRow(VLatest.ExtendedQueryTagString.TagKey));
-                }
-
-                return keys;
-            }
-            catch (SqlException ex)
-            {
-                throw ex.Number switch
-                {
-                    SqlErrorCodes.Conflict => ex.State == 1
-                        ? new ExtendedQueryTagsExceedsMaxAllowedCountException(maxAllowedCount)
-                        : new ExtendedQueryTagsAlreadyExistsException(),
-                    _ => new DataStoreException(ex),
-                };
-            }
-        }
-
-        public override async Task DeleteExtendedQueryTagAsync(string tagPath, string vr, bool force = false, CancellationToken cancellationToken = default)
-        {
-            using (SqlConnectionWrapper sqlConnectionWrapper = await ConnectionWrapperFactory.ObtainSqlConnectionWrapperAsync(cancellationToken))
-            using (SqlCommandWrapper sqlCommandWrapper = sqlConnectionWrapper.CreateSqlCommand())
-            {
-                VLatest.DeleteExtendedQueryTag.PopulateCommand(
-                    sqlCommandWrapper,
-                    tagPath,
-                    (byte)ExtendedQueryTagLimit.ExtendedQueryTagVRAndDataTypeMapping[vr],
-                    force);
-
-                try
-                {
-                    await sqlCommandWrapper.ExecuteNonQueryAsync(cancellationToken);
-                }
-                catch (SqlException ex)
-                {
-                    switch (ex.Number)
-                    {
-                        case SqlErrorCodes.NotFound:
-                            throw new ExtendedQueryTagNotFoundException(
-                                string.Format(CultureInfo.InvariantCulture, DicomSqlServerResource.ExtendedQueryTagNotFound, tagPath));
-                        case SqlErrorCodes.PreconditionFailed:
-                            throw new ExtendedQueryTagBusyException(
-                                string.Format(CultureInfo.InvariantCulture, DicomSqlServerResource.ExtendedQueryTagIsBusy, tagPath));
-                        default:
-                            throw new DataStoreException(ex);
-                    }
-                }
-            }
-        }
-    }
-}
