+++ src/Microsoft.Health.Dicom.SqlServer/Features/Schema/SchemaVersion.cs
@@ -14,6 +14,5 @@ namespace Microsoft.Health.Dicom.SqlServer.Features.Schema
         V1 = 1,
         V2 = 2,
         V3 = 3,
-        V4 = 4,
     }
 }
