+++ src/Microsoft.Health.Dicom.SqlServer/Features/Schema/SchemaVersionConstants.cs
@@ -8,7 +8,7 @@ namespace Microsoft.Health.Dicom.SqlServer.Features.Schema
     public static class SchemaVersionConstants
     {
         public const int Min = (int)SchemaVersion.V1;
-        public const int Max = (int)SchemaVersion.V4;
+        public const int Max = (int)SchemaVersion.V3;
         public const int SupportExtendedQueryTagSchemaVersion = (int)SchemaVersion.V2;
     }
 }
