+++ src/Microsoft.Health.Dicom.SqlServer/Microsoft.Health.Dicom.SqlServer.csproj
@@ -60,12 +60,6 @@
     <EmbeddedResource Include="Features\Schema\Migrations\3.sql">
       <InputToImmutableSqlGenerator>true</InputToImmutableSqlGenerator>
       <InputToMutableSqlGenerator>true</InputToMutableSqlGenerator>
-      <MutableClassVersion>3</MutableClassVersion>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Features\Schema\Migrations\4.diff.sql" />
-    <EmbeddedResource Include="Features\Schema\Migrations\4.sql">
-      <InputToImmutableSqlGenerator>true</InputToImmutableSqlGenerator>
-      <InputToMutableSqlGenerator>true</InputToMutableSqlGenerator>
       <MutableClassVersion>Latest</MutableClassVersion>
     </EmbeddedResource>
   </ItemGroup>
